9001: WIP refactor to move addDocketEntryForSystemGeneratedOrder into a useCaseHelper for reuse [skip ci]
