+++ shared/src/business/useCases/serveCaseToIrs/serveCaseToIrsInteractor.test.js
@@ -1,7 +1,6 @@
 /* eslint-disable max-lines */
 const {
   addDocketEntryForPaymentStatus,
-  addDocketEntryForSystemGeneratedDocument,
   serveCaseToIrsInteractor,
 } = require('./serveCaseToIrsInteractor');
 
@@ -19,7 +18,6 @@ const {
   PARTY_TYPES,
   PAYMENT_STATUS,
   SERVICE_INDICATOR_TYPES,
-  SYSTEM_GENERATED_DOCUMENT_TYPES,
 } = require('../../entities/EntityConstants');
 const {
   docketClerkUser,
@@ -625,7 +623,7 @@ describe('serveCaseToIrsInteractor', () => {
     expect(applicationContext.getUtilities().uploadToS3).toHaveBeenCalled();
   });
 
-  it.skip('should generate an order, upload it to s3, and get todayPlus60 for orderForFilingFee', async () => {
+  it('should generate an order, upload it to s3, and remove brackets from orderContent for orderForFilingFee', async () => {
     mockCase = {
       ...MOCK_CASE,
       orderForFilingFee: true,
@@ -638,17 +636,16 @@ describe('serveCaseToIrsInteractor', () => {
     expect(applicationContext.getDocumentGenerators().order).toHaveBeenCalled();
     expect(applicationContext.getUtilities().uploadToS3).toHaveBeenCalled();
 
-    // call off use case helper addDocketEntryForSystemGeneratedDocument and verify that the oldOrderForFilingFee is set to todayPlus60
-    // is structured as expected without brackets
-    // testing that cloneDeep actually updates the content inside due to "frozen" constant
-    // can you check that options.clonedSystemDocument does not contain bracketed values?
-    // expect(
-    //   await addDocketEntryForSystemGeneratedDocument().mock.calls[0][0].options,
-    // ).toMatchObject({
-    //   clonedSystemDocument: {
-    //     content: expect.not.stringContaining('['),
-    //   },
-    // });
+    // this ensures orderContent is not wrapped in brackets and cloneDeep worked to replace constant's content
+    // anything else worth testing here? todayPlus60 functionality is being unit tested in DateHandler
+    expect(
+      await applicationContext.getUseCaseHelpers()
+        .addDocketEntryForSystemGeneratedOrder.mock.calls[0][0].options,
+    ).toMatchObject({
+      clonedSystemDocument: {
+        content: expect.not.stringContaining('['),
+      },
+    });
   });
 });
 
@@ -708,77 +705,77 @@ describe('addDocketEntryForPaymentStatus', () => {
   });
 });
 
-describe('addDocketEntryForSystemGeneratedDocument', () => {
-  let user;
+//todo: move these tests into the useCaseHelper test instead of here
+// describe.skip('addDocketEntryForSystemGeneratedOrder', () => {
+//   let user;
 
-  const { noticeOfAttachmentsInNatureOfEvidence, orderForFilingFee } =
-    SYSTEM_GENERATED_DOCUMENT_TYPES;
+//   const { noticeOfAttachmentsInNatureOfEvidence, orderForFilingFee } =
+//     SYSTEM_GENERATED_DOCUMENT_TYPES;
 
-  beforeEach(() => {
-    user = applicationContext.getCurrentUser();
-  });
+//   beforeEach(() => {
+//     user = applicationContext.getCurrentUser();
+//   });
 
-  it('should increase the docket entries and upload a generated pdf for noticeOfAttachments', async () => {
-    const caseEntity = new Case(
-      {
-        ...MOCK_CASE,
-      },
-      { applicationContext },
-    );
+//   it('should increase the docket entries and upload a generated pdf for noticeOfAttachments', async () => {
+//     const caseEntity = new Case(
+//       {
+//         ...MOCK_CASE,
+//       },
+//       { applicationContext },
+//     );
 
-    const newDocketEntriesFromNewCaseCount =
-      caseEntity.docketEntries.length + 1;
+//     const newDocketEntriesFromNewCaseCount =
+//       caseEntity.docketEntries.length + 1;
 
-    await addDocketEntryForSystemGeneratedDocument({
-      applicationContext,
-      caseEntity,
-      systemGeneratedDocument: noticeOfAttachmentsInNatureOfEvidence,
-      user,
-    });
+//     await addDocketEntryForSystemGeneratedOrder({
+//       applicationContext,
+//       caseEntity,
+//       systemGeneratedDocument: noticeOfAttachmentsInNatureOfEvidence,
+//       user,
+//     });
 
-    expect(caseEntity.docketEntries.length).toEqual(
-      newDocketEntriesFromNewCaseCount,
-    );
+//     expect(caseEntity.docketEntries.length).toEqual(
+//       newDocketEntriesFromNewCaseCount,
+//     );
 
-    expect(applicationContext.getDocumentGenerators().order).toHaveBeenCalled();
-    const passedInNoticeTitle =
-      applicationContext.getDocumentGenerators().order.mock.calls[0][0].data
-        .orderTitle;
-    expect(passedInNoticeTitle).toEqual(passedInNoticeTitle.toUpperCase()); //asserts that the passed in title was uppercase
+//     expect(applicationContext.getDocumentGenerators().order).toHaveBeenCalled();
+//     const passedInNoticeTitle =
+//       applicationContext.getDocumentGenerators().order.mock.calls[0][0].data
+//         .orderTitle;
+//     expect(passedInNoticeTitle).toEqual(passedInNoticeTitle.toUpperCase()); //asserts that the passed in title was uppercase
 
-    expect(applicationContext.getUtilities().uploadToS3).toHaveBeenCalled();
-    expect(
-      applicationContext.getPersistenceGateway().saveDocumentFromLambda,
-    ).toHaveBeenCalled();
-    // eslint-disable-next-line max-lines
-  });
+//     expect(applicationContext.getUtilities().uploadToS3).toHaveBeenCalled();
+//     expect(
+//       applicationContext.getPersistenceGateway().saveDocumentFromLambda,
+//     ).toHaveBeenCalled();
+//     // eslint-disable-next-line max-lines
+//   });
 
-  it('should increase the docket entries and upload a generated pdf for orderForFilingFee', async () => {
-    const caseEntity = new Case(
-      {
-        ...MOCK_CASE,
-      },
-      { applicationContext },
-    );
+//   it('should increase the docket entries and upload a generated pdf for orderForFilingFee', async () => {
+//     const caseEntity = new Case(
+//       {
+//         ...MOCK_CASE,
+//       },
+//       { applicationContext },
+//     );
 
-    const newDocketEntriesFromNewCaseCount =
-      caseEntity.docketEntries.length + 1;
+//     const newDocketEntriesFromNewCaseCount =
+//       caseEntity.docketEntries.length + 1;
 
-    await addDocketEntryForSystemGeneratedDocument({
-      applicationContext,
-      caseEntity,
-      systemGeneratedDocument: orderForFilingFee,
-      user,
-    });
+//     await addDocketEntryForSystemGeneratedOrder({
+//       applicationContext,
+//       caseEntity,
+//       systemGeneratedDocument: orderForFilingFee,
+//       user,
+//     });
 
-    expect(caseEntity.docketEntries.length).toEqual(
-      newDocketEntriesFromNewCaseCount,
-    );
+//     expect(caseEntity.docketEntries.length).toEqual(
+//       newDocketEntriesFromNewCaseCount,
+//     );
 
-    expect(applicationContext.getDocumentGenerators().order).toHaveBeenCalled();
+//     expect(applicationContext.getDocumentGenerators().order).toHaveBeenCalled();
 
-    expect(
-      applicationContext.getPersistenceGateway().saveDocumentFromLambda,
-    ).toHaveBeenCalled();
-  });
-});
+//     expect(
+//       applicationContext.getPersistenceGateway().saveDocumentFromLambda,
+//     ).toHaveBeenCalled();
+//   });
